Remove pagination button and add "fetch on scroll" functionality
The initial number of items displayed on the extension list page was fixed
and not all available spaces was used to display items filling the space.

This was confusing for users, indeed the pagination button were not visible
enough and the user wasn't looking for them since the presentation of the
items makes him think that all of them where already fetched.

This commit fulfill two roles: (1) it ensures the available space will be
used to display as much item as possible (See midas.slicerappstore.pageLimit() function)
and (2) it remove the pagination link by using a "fetch items on scroll" mechanism.

The "fetch on scroll" mechanism is implemented with the help of a jquery
plugin developed by "Anderson Ferminiano" and tweaked to allow an easier
integration within ajax based application. Improvements have been contributed
upstream and are waiting to be reviewed. See https://github.com/andferminiano/jquery-scroll-pagination/pull/6