Fix loading of extension when visibility API is not available
The "update of category count after filtering" introduced by commit 1c4004c
broke the loading of the page in browser not supporting the visibility api.
This impacted both Slicer 4.2.2-1 and regular browser.

Following commit 1c4004c, the function "categoriesLoaded" was both updating
the category cound and setting up the scroll pagination.
In case visibility api wasn't supported, it was calling "applyFilter.
In that particular case, considering that function "fetchCategories" was
called within "applyFilter" and that "applyFilter" was called by
"categoriesLoaded". This was leading to an infinite loop.

This commit move back the setup of the scroll pagination into the document
ready function.

This fixes Slicer issue 3088.
See http://www.na-mic.org/Bug/view.php?id=3088