ENH: Add search support (back-end)
Add search support to slicerappstore, so that a 'search' query parameter
is correctly propagated through to the actual database query (requires a
corresponding change to slicerpackages).

Note that this is only the back-end support for now, e.g. a URL with a
'search' query parameter manually added. Front-end support (i.e. adding
a new form input field to the page UI) will be done in a subsequent
commit.