ENH: Improve styling of other appstore filters
Apply similar styling (except the search icon) to the revision filter
field as the search field so that the two are consistent, and also apply
similar styling to the <select> filters (arch, os) for improved
appearance, especially in alternate system color schemes. Also limit the
width of the revision filter field, which both improves appearance
(otherwise the search filter field seems to be off by its lonesome self)
and conserves space.