ENH: Further improve widget styling
Tweak the styling of text input fields to produce a more subtle effect,
and more importantly, one that is more consistent between browsers (the
old effect was rather excessive in gecko). Also add margin specification
for similar reasons (webkit defaults to 2px, but gecko uses 0, which
resulted in the fields looking a little squashed together). Rearrange
shadows to be in the intended order, and add 'dummy' shadows to
non-hover so that transitions work correctly. Tweak text shadow on
buttons slightly. Remove border effects from <select>, leaving just the
shadow effects.