ENH: Improve styling of breadcrumbs
Tweak styling so that the breadcrumb navigation bar in an extension view
page does not look so out of place. In particular, set CSS 'overflow' on
the parent so that the bar is not painted outside of the parent element
(thus losing the nice rounded corners and producing a weird shadow) and
change the border color to match the parent's border color. Also, reduce
the margins a bit, giving the bar a slimmer appearance.