ENH: Allow suppressing breadcrumbs
Teach extension view page to honor a new query parameter 'breadcrumbs'
which, if set to 'none', hides the breadcrumb bar entirely. This is
desired for viewing the detail pages of installed extensions from within
Slicer, as it does not make sense to be able to navigate back to the
list of available extensions in this context.